requirements.txt
anthropic
chromadb
sentence-transformers
rich
python-frontmatter
